factor out more openai handlers
