+++ src/inspect_ai/model/_call_tools.py
@@ -1,5 +1,6 @@
 import asyncio
 import inspect
+import json
 import types
 from dataclasses import is_dataclass
 from logging import getLogger
@@ -21,6 +22,7 @@ from typing import (
     is_typeddict,
 )
 
+import yaml
 from jsonschema import Draft7Validator
 from pydantic import BaseModel
 
@@ -465,3 +467,56 @@ def truncate_tool_output(
         )
     else:
         return None
+
+
+def tool_parse_error_message(arguments: str, ex: Exception) -> str:
+    return f"Error parsing the following tool call arguments:\n\n{arguments}\n\nError details: {ex}"
+
+
+def parse_tool_call(
+    id: str, function: str, arguments: str, tools: list[ToolInfo]
+) -> ToolCall:
+    error: str | None = None
+    arguments_dict: dict[str, Any] = {}
+
+    def report_parse_error(ex: Exception) -> None:
+        nonlocal error
+        error = tool_parse_error_message(arguments, ex)
+        logger.info(error)
+
+    # if the arguments is a dict, then handle it with a plain json.loads
+    arguments = arguments.strip()
+    if arguments.startswith("{"):
+        try:
+            arguments_dict = json.loads(arguments)
+        except json.JSONDecodeError as ex:
+            report_parse_error(ex)
+
+    # otherwise parse it as yaml (which will pickup unquoted strings, numbers, and true/false)
+    # and then create a dict that maps it to the first function argument
+    else:
+        tool_info = next(
+            (
+                tool
+                for tool in tools
+                if tool.name == function and len(tool.parameters.properties) > 0
+            ),
+            None,
+        )
+        if tool_info:
+            param_names = list(tool_info.parameters.properties.keys())
+            try:
+                value = yaml.safe_load(arguments)
+                arguments_dict[param_names[0]] = value
+            except yaml.error.YAMLError:
+                # If the yaml parser fails, we treat it as a string argument.
+                arguments_dict[param_names[0]] = arguments
+
+    # return ToolCall with error payload
+    return ToolCall(
+        id=id,
+        function=function,
+        arguments=arguments_dict,
+        type="function",
+        parse_error=error,
+    )
