+++ src/inspect_ai/model/_providers/util/hf_handler.py
@@ -8,9 +8,9 @@ from typing_extensions import override
 from inspect_ai.tool._tool_call import ToolCall
 from inspect_ai.tool._tool_info import ToolInfo
 
+from ..._call_tools import parse_tool_call, tool_parse_error_message
 from ..._chat_message import ChatMessageAssistant
 from .chatapi import ChatAPIHandler
-from .util import parse_tool_call, tool_parse_error_message
 
 logger = getLogger(__name__)
